4.20-4.23, added tokenExtractor, fixed tests

+++ controllers/blogs.js
@@ -1,15 +1,7 @@
 const blogRouter = require('express').Router()
+const middleware = require('../utils/middleware')
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -17,6 +9,7 @@ blogRouter.get('/', async (request, response) => {
     .populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
+
 
 blogRouter.get('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
@@ -27,16 +20,10 @@ blogRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogRouter.post('/', async (request, response) => {
+
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
-
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-
-  const user = await User.findById(decodedToken.id)
-  // const { name, number } = body
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
@@ -52,40 +39,31 @@ blogRouter.post('/', async (request, response) => {
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
   response.status(201).json(savedBlog)
-
-  // Blog.find({ name: body.name }).then(blogs => {
-  //   let id = blogs.map(person => person.id)[0]
-
-  //   if (id) {
-  //     Blog.findByIdAndUpdate(id, { name, number }, { new: true, runValidators: true, context: 'query' })
-  //       .then(updatedBlog => {
-  //         response.json(updatedBlog)
-  //       })
-  //       .catch(error => next(error))
-  //   }
-  //   else if (!id) {
-  //     const person = new Blog ({
-  //       name: body.name,
-  //       number: body.number,
-  //     })
-
-  //     person.save().then(savedBlog => {
-  //       console.log(`Added ${person.name} number ${person.number} to phonebook`)
-  //       response.json(savedBlog)
-  //     })
-  //       .catch(error => next(error))
-  //   }
-  // })
 })
 
-blogRouter.delete('/:id', async (request, response) => {
+
+blogRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== request.user.id) {
+    return response.status(401).json({ error: 'only logged in users can delete their blogs' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
 
-blogRouter.put('/:id', async (request, response) => {
+
+blogRouter.put('/:id', middleware.userExtractor, async (request, response) => {
   const body = request.body
-  const { title, author, url, likes } = body
+  const { likes } = body
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== request.user.id) {
+    return response.status(401).json({ error: 'only logged in users can update their blogs' })
+  }
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, { likes }, { new: true, runValidators: true, context: 'query' })
   response.status(200).json(updatedBlog)
